revert ngtools/aot PR
