+++ generators/client/templates/angular/_tsconfig-aot.json
@@ -28,7 +28,7 @@
         "noImplicitAny": false,
         "suppressImplicitAnyIndexErrors": true,
         "outDir": "<%= DIST_DIR %>app",
-        "lib": ["dom", "es2015"],
+        "lib": ["es6", "dom"],
         "typeRoots": [
             "node_modules/@types"
         ]
@@ -36,7 +36,7 @@
 
     "files": [
         "<%= MAIN_SRC_DIR %>app/app.module.ts",
-        "<%= MAIN_SRC_DIR %>app/app.main-aot.ts"
+        "<%= MAIN_SRC_DIR %>app/app.main.ts"
     ],
 
     "angularCompilerOptions": {
