--- generators/client/templates/angular/src/main/webapp/app/_app.main-aot.ts
@@ -1,31 +1,0 @@
-<%#
- Copyright 2013-2017 the original author or authors from the JHipster project.
-
- This file is part of the JHipster project, see https://jhipster.github.io/
- for more information.
-
- Licensed under the Apache License, Version 2.0 (the "License");
- you may not use this file except in compliance with the License.
- You may obtain a copy of the License at
-
-      http://www.apache.org/licenses/LICENSE-2.0
-
- Unless required by applicable law or agreed to in writing, software
- distributed under the License is distributed on an "AS IS" BASIS,
- WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- See the License for the specific language governing permissions and
- limitations under the License.
--%>
-import { platformBrowser } from '@angular/platform-browser';
-import { ProdConfig } from './blocks/config/prod.config';
-import { <%=angular2AppName%>AppModuleNgFactory } from '../../../../<%= BUILD_DIR %>aot/src/main/webapp/app/app.module.ngfactory';
-
-ProdConfig();
-
-if (module['hot']) {
-    module['hot'].accept();
-}
-
-platformBrowser().bootstrapModuleFactory(<%=angular2AppName%>AppModuleNgFactory)
-.then((success) => console.log(`Application started`))
-.catch((err) => console.error(err));
