--- generators/client/templates/angular/src/main/webapp/app/_vendor-aot.ts
@@ -1,10 +1,0 @@
-// Angular
-import '@angular/platform-browser';
-import '@angular/core';
-import '@angular/common';
-import '@angular/http';
-import '@angular/router';
-import '@angular/forms';
-
-// RxJS
-import 'rxjs';
